(fix) useConcepts should use POST to search for concepts
This should fix useConcepts to work with a single request using the
newly-added ability to request concept references via a POST request,
which will require 3.1.0 or higher of the REST API.

I also removed the well-intentioned but broken attempt to "chunk" the
concept references. While the idea was to maximize the size of a "page",
there are other limiting factors like the overall URL size that the REST
API maximum results value is a poor proxy for (and it's intended use is
the number of results sent back in a page of responses, not how many
parameters should be sent).