(fix) Re-write some core hooks in the form-processor-factory to reduce churn
In the FormProcessorFactory, there's serious potential for render
cascades from the combinations of `useEffect()` and `setState()` that
then trigger additional `useEffect()` and `setState()` calls. This
commit switches to more aggressively caching some of these key hooks so
that there's less potential for these cascade to lead to catastrophic
reloading issues.