Reduce churn in processor context
This is basically done by splitting the processor context into a "static"
part that is derived from the form itself and a "dynamic" part that may
be updated by components downstream of the FormProcessorFactory. This
mostly should optimize things so that components that depend on the
processor context should mostly get stable objects, assuming they
correctly use the properties of the processor.